Features needed are now passed as bitmask. Also had some mess with libs - hope it works.

+++ libs/ifyAPI/src/pl/poznan/put/cs/ify/api/YEvent.java
@@ -1,5 +1,5 @@
 package pl.poznan.put.cs.ify.api;
 
 public abstract class YEvent {
-	public abstract int getId();
+	public abstract long getId();
 }
