+++ libs/ifyAPI/src/pl/poznan/put/cs/ify/api/features/YBatteryEvent.java
@@ -4,11 +4,11 @@ import pl.poznan.put.cs.ify.api.Y;
 import pl.poznan.put.cs.ify.api.YEvent;
 
 public class YBatteryEvent extends YEvent {
-	public static final int ID = Y.Battery;
+	public static final long ID = Y.Battery;
 	private int mLevel;
 
 	@Override
-	public int getId() {
+	public long getId() {
 		return ID;
 	}
